requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python
pillow
tqdm
tensorboard
mediapipe==0.10.14
lpips
insightface
onnxruntime